# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SJM Social Media Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  # USERS
...
  /groups:
    get:
      operationId: listGroups
      tags: [groups]
      summary: List groups
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Group"

    post:
      operationId: createGroup
      tags: [groups]
      summary: Create a group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
  /groups/{id}:
    get:
      operationId: getGroup
      tags: [groups]
      summary: Get group details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
  /groups/{id}/join:
    post:
      operationId: joinGroup
      tags: [groups]
      summary: Join a group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK

  # MARKETPLACE
  /marketplace/products:
    get:
      operationId: listProducts
      tags: [marketplace]
      summary: List marketplace products
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

    post:
      operationId: createProduct
      tags: [marketplace]
      summary: List a product for sale
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /marketplace/products/{id}:
    get:
      operationId: getProduct
      tags: [marketplace]
      summary: Get product details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found

  # SEARCH
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: Admin list all users
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, banned, all]
            default: all
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
  /admin/users/{id}/ban:
    post:
      operationId: banUser
      tags: [admin]
      summary: Ban a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BanInput"
      responses:
        "200":
          description: OK
...
  /admin/posts/{id}:
    delete:
      operationId: adminDeletePost
      tags: [admin]
      summary: Admin delete a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [fullName, username, email, password]
      properties:
        fullName:
          type: string
        username:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string

    # USER
    User:
      type: object
      required: [id, fullName, username, email, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        username:
          type: string
        email:
          type: string
        avatarUrl:
          type: ["string", "null"]
        coverUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        isVerified:
          type: boolean
        isAdmin:
          type: boolean
        isBanned:
          type: boolean
        followersCount:
          type: integer
        followingCount:
          type: integer
        postsCount:
          type: integer
        createdAt:
          type: string
...
    AdminUser:
      type: object
      required: [id, fullName, username, email, createdAt, isBanned, postsCount]
      properties:
        id:
          type: integer
        fullName:
          type: string
        username:
          type: string
        email:
          type: string
        avatarUrl:
          type: ["string", "null"]
        isVerified:
          type: boolean
        isAdmin:
          type: boolean
        isBanned:
          type: boolean
        postsCount:
          type: integer
        followersCount:
          type: integer
        createdAt:
          type: string
        banReason:
          type: ["string", "null"]

    # POST
...
    Group:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        coverUrl:
          type: ["string", "null"]
        isPrivate:
          type: boolean
        membersCount:
          type: integer
        postsCount:
          type: integer
        isMember:
          type: boolean
        createdAt:
          type: string
    GroupInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        isPrivate:
          type: boolean

    # MARKETPLACE
    Product:
      type: object
      required: [id, sellerId, title, price, createdAt]
      properties:
        id:
          type: integer
        sellerId:
          type: integer
        seller:
          $ref: "#/components/schemas/User"
        title:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        category:
          type: string
        imageUrls:
          type: array
          items:
            type: string
        condition:
          type: string
          enum: [new, like_new, good, fair, poor]
        location:
          type: ["string", "null"]
        isAvailable:
          type: boolean
        viewsCount:
          type: integer
        createdAt:
          type: string
    ProductInput:
      type: object
      required: [title, price, category]
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        category:
          type: string
        imageUrls:
          type: array
          items:
            type: string
        condition:
          type: string
          enum: [new, like_new, good, fair, poor]
        location:
          type: ["string", "null"]

    # SEARCH
...
    BanInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
...