# lib/api-zod/src/generated/types/groupInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SJM Social Media Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface GroupInput {
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  isPrivate?: boolean;
}