# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SJM Social Media Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductCondition } from "./productCondition";
import type { User } from "./user";

export interface Product {
  id: number;
  sellerId: number;
  seller?: User;
  title: string;
  /** @nullable */
  description?: string | null;
  price: number;
  category?: string;
  imageUrls?: string[];
  condition?: ProductCondition;
  /** @nullable */
  location?: string | null;
  isAvailable?: boolean;
  viewsCount?: number;
  createdAt: string;
}