# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SJM Social Media Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductInputCondition } from "./productInputCondition";

export interface ProductInput {
  title: string;
  /** @nullable */
  description?: string | null;
  price: number;
  category: string;
  imageUrls?: string[];
  condition?: ProductInputCondition;
  /** @nullable */
  location?: string | null;
}