# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SJM Social Media Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  fullName: string;
  username: string;
  email: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  coverUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  isVerified?: boolean;
  isAdmin?: boolean;
  isBanned?: boolean;
  followersCount?: number;
  followingCount?: number;
  postsCount?: number;
  createdAt: string;
}